Improved parsing scripts
This commit improves the parsing scripts by adding better commenting, and
making them easier to extend. In particular, the Analyzer class in
parse_event_logs now accepts a filter function that makes it easier to
filter out a particular kind of jobs (e.g., only the later jobs in an
experiment, only the jobs that do disk reads, etc.). The copy_logs file
also is now easier to use: the main function returns the names of the local
files, so we can write different scripts that use that script to copy the
files, and then use their own independent logic to extract information about the
relevant jobs.

This commit also splits the plot for the continuous monitor into two
separate plots: one that includes information about the number of running tasks
and monotasks, and a second that reports utilization.