Simplification and cleanup of scripts.
This change makes a few simplifying changes to the parsing scripts:
(1) The job.py, stage.py, and task.py files are now part of this repository,
rather than needing to be read from a separate folder / repository. Those files
were copied from the monotasks branch of this repository:
https://github.com/kayousterhout/trace-analysis and then stripped to eliminate
functionality specific to the measurement work they were originally written for.

(2) The parse_event_logs script has also been stripped to eliminate all of the
functionality specific to the previous measurement work.

(3) All logging has been moved to use the logging library, so that it's easier to
turn on / off logging, as needed.