Plot number of monotasks on each disk.
This commit updates the continuous monitor parsing to parse the
number of running and queued monotasks on each disk, and then to
include this information on the disk utilization plot. This commit
also updates the range of disk throughputs shown on the disk utilization
to go up to 150MBps, to account for the post-deadline improvement
in EC2 disk performance.