Make copy_logs.py use old subprocess API.
This commit modifies copy_logs.py so that it does not use the
subprocess.check_output() function, which is not present in the version
of Python that is installed by default on the m2.4xlarge EC2 machines
that we use. copy_logs.py has been refactored to use subprocess.Popen()
instead.