Make plot_continuous_monitor parsing more robust
Previously, plot_continuous_monitor.py parsed the disk utilization logs
by hardcoding which index to use to extract logs for each disk. This
commit removes these hardcoded indices and changes
plot_continuous_monitor.py to instead search for the logs corresponding
to each disk. This is more robust in the event that the disks are logged
in a different order than is expected.