Added instrumentation for local macrotasks
This commit adds functionality to the continuous monitor parsing
to parse and graph the number of local macrotasks separately from
the total number of macrotasks that have monotasks running on a
particular executor. This is useful now that the count of running
macrotasks includes remote macrotasks that have monotasks running
on the machine (e.g., to fetch shuffle data).