Added new ideal time calculation based on utilization.
This commit adds a new method to stage.py to calculate the ideal
runtime of a stage based on the utilization while tasks were running,
rather than on the runtime of each type of monotask. This is useful
because it makes it possible to calculate the ideal runtime even
for regular Spark jobs.