Improvements to parse_vary_num_tasks.py script.
This commit makes the following improvements to the parse_vary_num_tasks.py
script:
-It now outputs an additional file, actual_runtimes_map_reduce.pdf, that
shows the actual and ideal runtimes separately for the map and reduce stages,
which can be useful for determining what's causing the ideal runtime to diverge
from the actual runtime.
-Copying the files back from the cluster is optional; if the user doesn't
specify a master name etc., the script will attempt to use existing files.
This is useful to re-parse data that's already been copied over. Enabling this
change required changing the order of the parameters to parse_vary_num_tasks.py.