Updated scripts so they work for the millennium cluster.
This commit adds an optional last parameter that can be used to
specify a username to log into the cluster (it defaults to
root, which is correct for the EC2 instances).

It also only adds "source /root/.bash_profile" when ssh'ing
from an EC2 cluster.

To make the disk parsing work properly, this commit changes
the parsing of disks to automatically detect what disks were
in used, and only plot those.  Some of the scripts will
incorrectly label the millennium disks (using xvda/xvdb instead
of sda1/sdb) but they will be correctly plotted. As part of
that change, this commit removes the file that plots the
information for all of the disks, because it's not as useful
as the per-disk plots (and contains the same information as
those plots), and it's more challenging to make it work
correctly with the automatic detection of which disks were
used.