Add throughputs to stage and job metrics
Modified the classes in metrics.py to calculate utilization and
effective throughput. Added a new metrics class: CpuMetrics. We cannot
calculate the disk read and disk write throughputs separately because
the disk counters that track read/write access time are not updated
live; instead, they are only updated when an access completes, which
leads to inaccuracies.