Improved metrics parsing.
Changed the utilization and effective disk throughput to be methods
rather than instance variables to avoid intermediate divide-by-zero
issues when tasks don't read any data, and also added explicit
handling for the case where the throughput is 0.

Added the actual job runtime to the ideal runtimes file, to make
it easier for us to check how close we are to ideal.