Cleaned up ideal time calculations.
This commit removes some now-redundant methods for calculating the
ideal time of a job so there's just one ideal time method for a stage.
That method sanity checks the job metrics about network and CPU time
with the metrics read from the OS counters, and then uses the job
metrics combined with the throughput inferred from the OS counters
to infer the ideal time.