Log warning rather than failing when sanity check fails.
There are quite a few cases when the sanity check fails (e.g., when
doing an early shuffle, or when the amount of data is small),
and it's hard to incorporate all of these cases into the sanity
checking code. Instead, this code changes the sanity check to log
a warning (rather than throwing an exception) when the sanity check
fails.