Write the distribution of map, reduce, and total times for sort jobs.
This commit changes the parse_sort.py script to output an additional set
of files that has the distribution of actual and ideal map, reduce, and
total runtimes. Each file includes the distribution for each experiment
that was in that parent directory, and the experiments are sorted by
the number of values per key.